# projecteBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Conecta Salud API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/alerts": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Llistar alertes",
                "description": "Retorna una col·lecció d'alertes amb filtres opcionals",
                "operationId": "4478ffd7354aa9ca6132fb9e74bdde0b",
                "parameters": [
                    {
                        "name": "operatorId",
                        "in": "query",
                        "description": "Filtrar per operador",
                        "required": false
                    },
                    {
                        "name": "patientId",
                        "in": "query",
                        "description": "Filtrar per pacient",
                        "required": false
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Mostrar alertes a partir d'una data",
                        "required": false
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "description": "Mostrar alertes fins a una data",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Llista d'alertes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Crear alerta",
                "description": "Crea una nova alerta",
                "operationId": "42b5297a4f4f281e294bf8339a0a80d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAlertRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alerta creada correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validació o de creació"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/alerts/{id}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Mostrar alerta",
                "description": "Retorna els detalls d'una alerta específica",
                "operationId": "48e65a8f03a14c0b36563c5284e726a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de l'alerta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalls de l'alerta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alerta no trobada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Actualitzar alerta",
                "description": "Actualitza una alerta existent",
                "operationId": "037adc62bb0d80e7c7a153cf2445e337",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de l'alerta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAlertRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alerta actualitzada correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error d'actualització"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Eliminar alerta",
                "description": "Elimina una alerta específica",
                "operationId": "6dea41ff5a0bbf10775b84b2275d50e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de l'alerta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Alerta eliminada correctament"
                    },
                    "400": {
                        "description": "Error en l'eliminació"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/calls/alerts/{id}": {
            "get": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Obtenir trucades per alerta",
                "description": "Retorna les trucades associades a una alerta específica",
                "operationId": "47681a7a5fc557743ae15937a5067068",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de l'alerta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trucades recuperades correctament",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OutgoingCallResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alerta no trobada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreAlertRequest": {
                "description": "Validació per a la creació d'una alerta",
                "required": [
                    "operatorId",
                    "patientId",
                    "type",
                    "isRecurring",
                    "date",
                    "time",
                    "title"
                ],
                "properties": {
                    "operatorId": {
                        "description": "ID de l'operador",
                        "type": "integer",
                        "example": 1
                    },
                    "patientId": {
                        "description": "ID del pacient",
                        "type": "integer",
                        "example": 1
                    },
                    "isActive": {
                        "description": "Indica si l'alerta està activa",
                        "type": "boolean",
                        "example": true
                    },
                    "type": {
                        "description": "Tipus d'alerta",
                        "type": "string",
                        "example": "emergency"
                    },
                    "isRecurring": {
                        "description": "Indica si l'alerta és recurrent",
                        "type": "boolean",
                        "example": false
                    },
                    "date": {
                        "description": "Data de l'alerta",
                        "type": "string",
                        "format": "date",
                        "example": "2023-10-01"
                    },
                    "endDate": {
                        "description": "Data de finalització de l'alerta",
                        "type": "string",
                        "format": "date",
                        "example": "2023-10-02"
                    },
                    "time": {
                        "description": "Hora de l'alerta",
                        "type": "string",
                        "format": "time",
                        "example": "14:30:00"
                    },
                    "title": {
                        "description": "Títol de l'alerta",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Reunió amb el metge"
                    },
                    "dayOfWeek": {
                        "description": "Dies de la setmana per a l'alerta recurrent",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Monday",
                            "Wednesday"
                        ]
                    },
                    "description": {
                        "description": "Descripció de l'alerta",
                        "type": "string",
                        "example": "Reunió de seguiment amb el metge"
                    }
                },
                "type": "object"
            },
...
            "AlertResource": {
                "description": "Esquema del recurs Alert",
                "properties": {
                    "id": {
                        "description": "Identificador de l'alerta",
                        "type": "integer",
                        "example": 1
                    },
                    "operatorId": {
                        "description": "Identificador de l'operador",
                        "type": "integer",
                        "example": 2
                    },
                    "operator": {
                        "$ref": "#/components/schemas/OperatorResource"
                    },
                    "patientId": {
                        "description": "Identificador del pacient",
                        "type": "integer",
                        "example": 5
                    },
                    "patient": {
                        "$ref": "#/components/schemas/PatientSimpleResource"
                    },
                    "zoneId": {
                        "description": "Identificador de la zona",
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "zone": {
                        "$ref": "#/components/schemas/ZoneResource"
                    },
                    "isActive": {
                        "description": "Estat actiu de l'alerta",
                        "type": "boolean",
                        "example": true
                    },
                    "type": {
                        "description": "Tipus d'alerta",
                        "type": "string",
                        "example": "mèdica"
                    },
                    "isRecurring": {
                        "description": "Indica si l'alerta és recurrent",
                        "type": "boolean",
                        "example": false
                    },
                    "date": {
                        "description": "Data de l'alerta",
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-20"
                    },
                    "endDate": {
                        "description": "Data de finalització de l'alerta (si és recurrent)",
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-20",
                        "nullable": true
                    },
                    "time": {
                        "description": "Hora de l'alerta",
                        "type": "string",
                        "format": "time",
                        "example": "14:30:00"
                    },
                    "dayOfWeek": {
                        "description": "Dies de la setmana per a alertes recurrents",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "dilluns",
                                "dimarts",
                                "dimecres",
                                "dijous",
                                "divendres",
                                "dissabte",
                                "diumenge"
                            ]
                        },
                        "example": [
                            "dilluns",
                            "dimecres",
                            "divendres"
                        ]
                    },
                    "description": {
                        "description": "Descripció de l'alerta",
                        "type": "string",
                        "example": "Visita mèdica programada",
                        "nullable": true
                    },
                    "title": {
                        "description": "Títol de l'alerta",
                        "type": "string",
                        "example": "Control de tensió"
                    }
                },
                "type": "object"
            },
...
            "LanguageResource": {
                "description": "Esquema del recurs Language",
                "properties": {
                    "id": {
                        "description": "Identificador de l'idioma",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de l'idioma",
                        "type": "string",
                        "example": "Català"
                    }
                },
                "type": "object"
            },
            "OperatorResource": {
                "description": "Esquema del recurs Operator",
                "properties": {
                    "id": {
                        "description": "Identificador de l'operador",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de l'operador",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Correu electrònic de l'operador",
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "description": "Rol de l'operador",
                        "type": "string",
                        "example": "admin"
                    },
                    "surnames": {
                        "description": "Cognoms de l'operador",
                        "type": "string",
                        "example": "Doe"
                    },
                    "phone": {
                        "description": "Telèfon de l'operador",
                        "type": "string",
                        "example": "123456789"
                    },
                    "hireDate": {
                        "description": "Data de contractació de l'operador",
                        "type": "string",
                        "format": "date",
                        "example": "2022-01-01"
                    },
                    "terminationDate": {
                        "description": "Data de finalització de l'operador",
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01",
                        "nullable": true
                    },
                    "zones": {
                        "description": "Zones associades a l'operador",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ZoneResource"
                        }
                    },
                    "patients": {
                        "description": "Pacients associats a l'operador",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PatientResource"
                        }
                    },
                    "languages": {
                        "description": "Idiomes parlats per l'operador",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LanguageResource"
                        }
                    }
                },
                "type": "object"
            },
            "OutgoingCallResource": {
                "description": "Esquema del recurs OutgoingCall",
                "properties": {
                    "callId": {
                        "description": "Identificador de la trucada",
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "description": "Tipus de trucada",
                        "type": "string",
                        "example": "emergència"
                    },
                    "alert": {
                        "$ref": "#/components/schemas/AlertResource"
                    }
                },
                "type": "object"
            },
            "PatientSimpleResource": {
                "description": "Informació simplificada del pacient",
                "properties": {
                    "id": {
                        "description": "Identificador del pacient",
                        "type": "integer",
                        "example": 12
                    },
                    "fullName": {
                        "description": "Nom complet del pacient",
                        "type": "string",
                        "example": "Ana García López"
                    },
                    "birthDate": {
                        "description": "Data de naixement",
                        "type": "string",
                        "format": "date",
                        "example": "1985-07-22"
                    },
                    "fullAddress": {
                        "description": "Adreça completa",
                        "type": "string",
                        "example": "Carrer Major, 45, 2n A, Alcoi, 03801"
                    },
                    "dni": {
                        "description": "DNI del pacient",
                        "type": "string",
                        "example": "12345678A"
                    },
                    "healthCardNumber": {
                        "description": "Número de la targeta sanitària",
                        "type": "string",
                        "example": "SIP87654321"
                    },
                    "phone": {
                        "description": "Número de telèfon",
                        "type": "string",
                        "example": "+34 654 321 987"
                    },
                    "email": {
                        "description": "Correu electrònic",
                        "type": "string",
                        "format": "email",
                        "example": "ana.garcia@example.com"
                    },
                    "personalFamilySituation": {
                        "description": "Situació personal i familiar",
                        "type": "string",
                        "example": "Viu sola, té dos fills que viuen a prop."
                    },
                    "healthSituation": {
                        "description": "Situació de salut",
                        "type": "string",
                        "example": "Diabetis tipus 2 controlada amb medicació."
                    },
                    "housingSituation": {
                        "description": "Situació d'habitatge",
                        "type": "string",
                        "example": "Habitatge propi amb ascensor."
                    },
                    "personalAutonomy": {
                        "description": "Autonomia personal",
                        "type": "string",
                        "example": "Autònoma per a activitats bàsiques, necessita ajuda per a la compra."
                    },
                    "economicSituation": {
                        "description": "Situació econòmica",
                        "type": "string",
                        "example": "Pensió contributiva de 800€ mensuals."
                    }
                },
                "type": "object"
            },
...
            "ZoneResource": {
                "description": "Esquema del recurs Zone",
                "properties": {
                    "id": {
                        "description": "Identificador de la zona",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de la zona",
                        "type": "string",
                        "example": "Zona 1"
                    },
                    "status": {
                        "description": "Estat de la zona",
                        "type": "string",
                        "example": "activa"
                    },
                    "created_at": {
                        "description": "Data de creació",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "description": "Data d'actualització",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-02T00:00:00Z"
                    },
                    "operators": {
                        "description": "Llista d'operadors associats a la zona",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OperatorResource"
                        }
                    },
                    "patients": {
                        "description": "Llista de pacients associats a la zona",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PatientResource"
                        }
                    },
                    "alerts": {
                        "description": "Llista d'alertes associades a la zona",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AlertResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}